list and data reogranized
