+++ client/src/api/hooks/threads/data/useThreadBasicData.ts
@@ -1,0 +1,27 @@
+import { useQuery } from "@tanstack/react-query";
+import { ForumThread } from "@/lib/types";
+import { fetchThreadById } from "../../../queries/thread";
+
+interface UseThreadBasicDataOptions {
+  threadId: string;
+  userId?: string;
+}
+
+export function useThreadBasicData({ threadId, userId }: UseThreadBasicDataOptions) {
+  // Fetch thread data
+  const {
+    data: thread,
+    isLoading,
+    error,
+  } = useQuery<ForumThread>({
+    queryKey: [`/api/threads/id/${threadId}`, userId],
+    queryFn: () => fetchThreadById(threadId, userId),
+    enabled: !!threadId,
+  });
+
+  return {
+    thread,
+    isLoading,
+    error,
+  };
+}
